refactor(space): nest under library module
Spaces are intrinsically library-scoped (third top-level folder
alongside `lib/` and `runtime/`), so they belong inside the library
subsystem rather than at the crate root.

- `core/src/space/` → `core/src/library/space/`
- `core/src/library/mod.rs` adds `pub mod space` + re-exports `Spaces`
- `core/src/lib.rs` drops the top-level `mod space;` and pulls
  `Spaces` from `library::Spaces`
- The MCP `SpacesTool` updates its import to `crate::library::Spaces`

`SpaceId` and `AuthResource::Space` stay where they are
(`primitives` and `auth/resource.rs` respectively) because they're
referenced by other top-level modules (sandbox mode, authz checks).
